feat(FZF): better fzf integration
The app now has commands to make it easier to integrate with fzf and
similar. It also has a new docs/FZF-INTEGRATION.md which gives you hint
on how to integrate notr with fish/bash/zsh.
A problem was also fixed where the title was added at the beginning of
the note every time it was edited.